requirements.txt
python-telegram-bot
sqlmodel
matplotlib
pandas
